requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
attrs==25.1.0
backoff==2.2.1
bert-score==0.3.13
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
datasets==3.1.0
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
filelock==3.16.1
fonttools==4.55.7
frozenlist==1.5.0
fsspec==2023.10.0
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.7
Levenshtein==0.25.1
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.1
nltk==3.9.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
openai==1.60.2
packaging==24.2
pandas==2.0.3
parrot @ git+https://github.com/PrithivirajDamodaran/Parrot_Paraphraser.git@03084c54b64019ba5fa0b620b9c70ad81123e458
pillow==10.4.0
pkg_resources==0.0.0
propcache==0.2.0
protobuf==5.29.3
pyarrow==17.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
pytz==2024.2
PyYAML==6.0.2
rapidfuzz==3.9.7
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==3.2.1
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.4.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
